Updating the non-autogenerated (manual code)
Althought most of the parsing code is automatic a majority a bunch of
stuff have to be manually written. Although some of it could be
generated automatically (like dependencies) and boiler plate code for
visitors (run method) we will be some manual work till we reach some
level of maturity.