Adding 2d mouse navigation.
The mouse navigation is handled by 2d-mouse class in device.lisp. All
devices should return a transformation matrix. The entire mechanism and
state is internal to the class. All other devices (be it multitouch etc)
should have this type of interface henceforth.

The translation functionality is not complete. We only have a working
rotation implementation. We need to implement pan and zoom as well.