Adding Screen and Head physical devices.
Other than the 2d-mouse we are now adding Head (tracking) and
Screen (display) device classes and corresponding run methods.

This will enable us to do some VR work with cl-3d. The default single
screen configuration defaults to something that can be used on a deskop
with a 45 degree FOV. This can remain as it is for a desktop.

Head is currently identiry but will enable use to do some head tracked
VR once connected to a tracking device.