Adding VR mathematics for Projection and ModelView
VR typically takes into account the coordinates of the tracked head
with respect to a base and screen coordinates wrt the same base. This
helps in aligning the virtual work with the real world.

CAVE's TiledWall and other VR setups can be used or atleast we have the
capability to use them.