Adding quadbuffer (crystal eye) stereo.
Adding code to do quadbuffer stereo.

In the future we should have stereo window which will do this for
us. Also the display configuration and the device management should be a
part of the window just like it is for any other system. The window
provides IO (display/stereo/mouse/keyboard etd). The user does not deal
with setting up Model-view or projection or window. The scene that is
specified will be taken by the VRWindow and rendered accordingly.

For now the stereo mode is off.