Moving Global Variables into device object
Collecting all global variables in to device class. The device class
keeps a collection of all the device in the real world. Currently we
have screen, head and 2d-mouse.

The device is accessble form the scene. This will perhaps be moved in
the future cause it logically devices are at the same level as the scene
and not contained within it.

Devices could still further be decomposed into input and output devices.