Defining *Scene* global and VR class
To decouple the scene we are making a variable Scene to hold global
parameter. We also move all the windowing code to VR class.

The VR class contains the scene and device parmeters. However currently
we are leaving the *Scene* as a global variable since the VR class has a
broader role which is not currently defined.