Moving rendering logic.
The VR window takes in the head and the screen to compute the projection
matrix. We were previously computing two matrices namely model-view and
projection seperately. However for the sake of simplicity they can be
seena as one projection matrix.

The scene is thus free to now use the model-view matrix and do things
totally independent of how the configuration is setup in the VR class.