Movin all files in packages to X3D
We want to clean up the main directory and divide all modules into
subdirectories. However to maintain all the things that are being tested
by the test cased we are only moving files which are in the package and
not all the files which are autogenerated. This gives us an idea about
what test cases are still not available.